+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Support ecoinvent 3.5
 * Drop unused and outdated ecoinvent 31 biosphere flows
 * Add CPC classification to ecoinvent imports
+* Make importers play better with the Activity Browser
 
 ## 0.6 (2018-05-31)